Move the heartbeat endpoint to its own app
A scraper_api.apps folder has been created to hold all Django apps of
this project and the heartbeat has been moved here as a new Django app
